+++ git-ref/tests/transaction/mod.rs
@@ -3,7 +3,7 @@ mod refedit_ext {
 
     use bstr::{BString, ByteSlice};
     use git_ref::{
-        transaction::{Change, Create, RefEdit, RefEditsExt, RefLog},
+        transaction::{Change, PreviousValue, RefEdit, RefEditsExt, RefLog},
         PartialNameRef, Target,
     };
 
@@ -109,7 +109,7 @@ mod refedit_ext {
             RefEdit {
                 change: Change::Update {
                     log: Default::default(),
-                    previous: Create::Only,
+                    previous: PreviousValue::MustNotExist,
                     new: Target::Symbolic("refs/heads/main".try_into()?),
                 },
                 name: "HEAD".try_into()?,
@@ -131,7 +131,7 @@ mod refedit_ext {
                 RefEdit {
                     change: Change::Update {
                         log: Default::default(),
-                        previous: Create::Only,
+                        previous: PreviousValue::MustNotExist,
                         new: Target::Symbolic("refs/namespaces/foo/refs/heads/main".try_into()?),
                     },
                     name: "refs/namespaces/foo/HEAD".try_into()?,
@@ -148,7 +148,7 @@ mod refedit_ext {
 
         use git_hash::ObjectId;
         use git_ref::{
-            transaction::{Change, Create, LogChange, RefEdit, RefEditsExt, RefLog},
+            transaction::{Change, LogChange, PreviousValue, RefEdit, RefEditsExt, RefLog},
             FullNameRef, PartialNameRef, Target,
         };
         use git_testtools::hex_to_id;
@@ -249,7 +249,7 @@ mod refedit_ext {
                 },
                 RefEdit {
                     change: Change::Update {
-                        previous: Create::Only,
+                        previous: PreviousValue::MustNotExist,
                         log: LogChange {
                             mode: RefLog::AndReference,
                             force_create_reflog: true,
@@ -323,7 +323,7 @@ mod refedit_ext {
                 },
                 RefEdit {
                     change: Change::Update {
-                        previous: Create::Only,
+                        previous: PreviousValue::MustNotExist,
                         log: log.clone(),
                         new: Target::Peeled(ObjectId::null_sha1()),
                     },
@@ -359,7 +359,7 @@ mod refedit_ext {
                     },
                     RefEdit {
                         change: Change::Update {
-                            previous: Create::OrUpdate { previous: None },
+                            previous: PreviousValue::Any,
                             log: log_only.clone(),
                             new: Target::Peeled(ObjectId::null_sha1()),
                         },
@@ -376,7 +376,7 @@ mod refedit_ext {
                     },
                     RefEdit {
                         change: Change::Update {
-                            previous: Create::OrUpdate { previous: None },
+                            previous: PreviousValue::Any,
                             log: log_only,
                             new: Target::Peeled(ObjectId::null_sha1()),
                         },
@@ -393,7 +393,7 @@ mod refedit_ext {
                     },
                     RefEdit {
                         change: Change::Update {
-                            previous: Create::Only,
+                            previous: PreviousValue::MustNotExist,
                             log,
                             new: Target::Peeled(ObjectId::null_sha1()),
                         },