+++ git-repository/src/commit.rs
@@ -1,5 +1,6 @@
-use bstr::{BStr, BString, ByteSlice, ByteVec};
 use std::borrow::Cow;
+
+use bstr::{BStr, BString, ByteSlice, ByteVec};
 
 /// An empty array of a type usable with the `git::easy` API to help declaring no parents should be used
 pub const NO_PARENT_IDS: [git_hash::ObjectId; 0] = [];